@@ -3,12 +3,12 @@ package crane
 import "fmt"
 
 type NetworkParameters struct {
-	RawAlias       interface{} `json:"alias" yaml:"alias"`
-	RawAliases     interface{} `json:"aliases" yaml:"aliases"`
-	RawIp          string      `json:"ip" yaml:"ip"`
-	RawIpv4Address string      `json:"ipv4_address" yaml:"ipv4_address"`
-	RawIp6         string      `json:"ip6" yaml:"ip6"`
-	RawIpv6Address string      `json:"ipv6_address" yaml:"ipv6_address"`
+	RawAlias       any    `json:"alias" yaml:"alias"`
+	RawAliases     any    `json:"aliases" yaml:"aliases"`
+	RawIp          string `json:"ip" yaml:"ip"`
+	RawIpv4Address string `json:"ipv4_address" yaml:"ipv4_address"`
+	RawIp6         string `json:"ip6" yaml:"ip6"`
+	RawIpv6Address string `json:"ipv6_address" yaml:"ipv6_address"`
 }
 
 // If aliases are not defined in the config,
@@ -26,7 +26,7 @@ func (n NetworkParameters) Alias(containerName string) []string {
 		aliases = append(aliases, containerName)
 	} else {
 		switch concreteValue := rawAliases.(type) {
-		case []interface{}:
+		case []any:
 			for _, v := range concreteValue {
 				aliases = append(aliases, expandEnv(v.(string)))
 			}
@@ -49,14 +49,14 @@ func (n NetworkParameters) Ip6() string {
 	return expandEnv(n.RawIpv6Address)
 }
 
-func createNetworkParemetersFromMap(val map[string]interface{}) NetworkParameters {
+func createNetworkParemetersFromMap(val map[string]any) NetworkParameters {
 	var params NetworkParameters
 	if val == nil {
 		params = NetworkParameters{}
 	} else {
 		for k, v := range val {
 			if k == "alias" || k == "aliases" {
-				params.RawAlias = v.([]interface{})
+				params.RawAlias = v.([]any)
 			} else if k == "ip" || k == "ipv4_address" {
 				params.RawIp = v.(string)
 			} else if k == "ip6" || k == "ipv6_address" {
